normal_atom class now has static_cast<> for the abs function in order to avoid a commmon warning of comparison between signed and unsigned integer expressions

+++ chemistry/general/atoms/normal_atom.hpp
@@ -20,7 +20,7 @@ namespace msci
 		public:
 			normal_atom(int new_ionic_charge = 0,int new_neutrons_number = T::mass_number) : atom_crtp<T>(),mass_number(new_neutrons_number),ionic_charge(new_ionic_charge)
 			{
-				if (new_ionic_charge > atom_crtp<T>::get_z())
+				if (static_cast<unsigned int>(new_ionic_charge) > atom_crtp<T>::get_z())
 				{
 					throw "ionic charge is greater than z";
 				}
@@ -33,7 +33,7 @@ namespace msci
 
 			virtual bool is_factible() const
 			{
-				if (get_ionic_charge() > atom_crtp<T>::get_z())
+				if (static_cast<unsigned int>(get_ionic_charge()) > atom_crtp<T>::get_z())
 				{
 					return false;
 				}
@@ -41,14 +41,14 @@ namespace msci
 				{
 					if (atom::is_atom_specimen(atom_symbol::H) or atom::is_atom_specimen(atom_symbol::He))
 					{
-						if ((get_ionic_charge() < 0) and (std::abs(get_ionic_charge()) > (2 - atom_crtp<T>::get_z())))
+						if ((get_ionic_charge() < 0) and (static_cast<unsigned int>(std::abs(get_ionic_charge())) > (2 - atom_crtp<T>::get_z())))
 						{
 							return false;
 						}
 					}
 					else
 					{
-						if ((get_ionic_charge() < 0) and (std::abs(get_ionic_charge()) > (8 - atom_crtp<T>::get_z())))
+						if ((get_ionic_charge() < 0) and (static_cast<unsigned int>(std::abs(get_ionic_charge())) > (8 - atom_crtp<T>::get_z())))
 						{
 							return false;
 						}
